% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{compute_arrangement}
\alias{compute_arrangement}
\title{Compute Arrangement Using Jonker & Volgenant algorithm}
\usage{
compute_arrangement(mydata, macro_x, macro_y, faceter, n_row, n_col)
}
\arguments{
\item{mydata}{A data frame which contains the {macro_x, macro_y, faceter} columns}

\item{macro_x}{the name of a column which shall be used to arrange facets horizontally}

\item{macro_y}{the name of a column which shall be used to arrange facets vertically}

\item{faceter}{the column name defining faceting groups}

\item{n_row, n_col}{Number of rows and columns.}
}
\description{
Compute Arrangement Using Jonker & Volgenant algorithm
}
