% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facetwarp.R
\name{facet_warp}
\alias{facet_warp}
\title{Arrange Facets for your ggplot object}
\usage{
facet_warp(
  facets,
  macro_x,
  macro_y,
  nrow = NULL,
  ncol = NULL,
  strip.position = "top",
  labeller = "label_value",
  drop = TRUE
)
}
\arguments{
\item{facets}{A set of variables or expressions quoted by \code{\link[ggplot2:vars]{vars()}}
and defining faceting groups on the rows or columns dimension.
The variables can be named (the names are passed to \code{labeller}).

For compatibility with the classic interface, can also be a
formula or character vector. Use either a one sided formula, \code{~a + b},
or a character vector, \code{c("a", "b")}.}

\item{macro_x}{the name of a column which shall be used to arrange facets horizontally}

\item{macro_y}{the name of a column which shall be used to arrange facets vertically}

\item{nrow, ncol}{Number of rows and columns.}

\item{strip.position}{By default, the labels are displayed on the top of
the plot. Using \code{strip.position} it is possible to place the labels on
either of the four sides by setting \code{strip.position = c("top",
  "bottom", "left", "right")}}

\item{labeller}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with \code{vars(cyl, am)}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link[ggplot2:labeller]{labeller()}}. You can use different labeling
functions for different kind of labels, for example use \code{\link[ggplot2:label_parsed]{label_parsed()}} for
formatting facet labels. \code{\link[ggplot2:label_value]{label_value()}} is used by default,
check it for more details and pointers to other options.}

\item{drop}{If \code{TRUE}, the default, all factor levels not used in the
data will automatically be dropped. If \code{FALSE}, all factor levels
will be shown, regardless of whether or not they appear in the data.}
}
\description{
Arrange Facets for your ggplot object
}
\examples{
ggplot(iris)+
   geom_point(aes(x=Petal.Width, y=Petal.Length))+
   facet_warp(vars(Species), macro_x='Sepal.Width', macro_y='Sepal.Length', nrow = 2)
}
